+++ cpp/src/arrow/compute/kernels/aggregate_basic.cc
@@ -16,9 +16,9 @@
 // under the License.
 
 #include "arrow/compute/api_aggregate.h"
-#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/kernels/aggregate_basic_internal.h"
 #include "arrow/compute/kernels/aggregate_internal.h"
+#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/kernels/common_internal.h"
 #include "arrow/compute/kernels/util_internal.h"
 #include "arrow/util/cpu_info.h"
