+++ cpp/src/arrow/compute/kernels/copy_data_internal.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "arrow/compute/codegen_internal.h"
+#include "arrow/compute/kernels/codegen_internal.h"
 
 namespace arrow {
 namespace compute {
