+++ cpp/src/arrow/compute/kernels/scalar_cast_string.cc
@@ -20,8 +20,8 @@
 
 #include "arrow/array/array_base.h"
 #include "arrow/array/builder_binary.h"
-#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/kernels/base_arithmetic_internal.h"
+#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/kernels/common_internal.h"
 #include "arrow/compute/kernels/scalar_cast_internal.h"
 #include "arrow/compute/kernels/temporal_internal.h"
