+++ cpp/src/arrow/compute/kernels/scalar_cast_test.cc
@@ -44,8 +44,8 @@
 
 #include "arrow/compute/api_vector.h"
 #include "arrow/compute/cast.h"
-#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/kernel.h"
+#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/kernels/test_util_internal.h"
 
 namespace arrow {
