+++ cpp/src/arrow/compute/kernels/vector_selection_internal.cc
@@ -27,9 +27,9 @@
 #include "arrow/buffer_builder.h"
 #include "arrow/chunked_array.h"
 #include "arrow/compute/api_vector.h"
-#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/function.h"
 #include "arrow/compute/kernel.h"
+#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/kernels/vector_selection_internal.h"
 #include "arrow/compute/registry.h"
 #include "arrow/type.h"
