+++ cpp/src/arrow/compute/kernels/vector_statistics.cc
@@ -22,10 +22,10 @@
 
 #include "arrow/compute/api_aggregate.h"
 #include "arrow/compute/api_vector.h"
-#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/exec.h"
 #include "arrow/compute/function.h"
 #include "arrow/compute/kernel.h"
+#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/registry.h"
 #include "arrow/result.h"
 #include "arrow/scalar.h"
